refactored code a bit
