+++ index.html
@@ -7,7 +7,7 @@
       href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap"
       rel="stylesheet"
     />
-    <script src="app.js" type="module"></script>
+    <script src="src/app.js" type="module"></script>
     <link rel="stylesheet" href="style.css" />
     <title>Who needs to share the board?</title>
   </head>
