+++ package.json
@@ -9,12 +9,17 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest"
   },
   "devDependencies": {
-    "vite": "^4.4.9"
+    "@vitest/ui": "^0.34.3",
+    "happy-dom": "^10.11.2",
+    "vite": "^4.4.9",
+    "vitest": "^0.34.3"
   },
   "dependencies": {
+    "copy-to-clipboard": "^3.3.3",
     "party-js": "^2.2.0"
   }
 }
