+++ src/teams.js
@@ -1,0 +1,30 @@
+export const teams = {
+    teamOne: [
+        "Michiel 1",
+        "Michiel 2",
+        "Jesse",
+        "Kevin",
+        "Frank",
+        "Ralph",
+        "Louise",
+        "Tarek",
+        "Ashwas",
+        "David",
+        "Bhaskara",
+        "Ivan",
+        "Martins",
+    ],
+
+    teamRocket: [
+        "Ivan",
+        "Maria",
+        "Sandhya",
+        "Dries",
+        "Nick",
+        "Tonny",
+        "Alex",
+        "Linh",
+        "Kenny",
+        "Liz",
+    ]
+};
