+++ style.css
@@ -3,17 +3,17 @@
     margin: 0;
     padding: 0;
     box-sizing: border-box;
-  }
+}
 
-  body {
+body {
     display: flex;
     justify-content: center;
     align-items: center;
     min-height: 100vh;
     background-color: #f3f3f3;
-  }
+}
 
-  button {
+button {
     background-color: red;
     color: white;
     font-size: 24px;
@@ -23,45 +23,45 @@
     border-radius: 10px;
     cursor: pointer;
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
-  }
+}
 
-  button:hover {
+button:hover {
     box-shadow: 0 0 15px rgba(0, 0, 0, 0.3);
-  }
+}
 
-  #chosen-name {
+#chosen-name {
     display: none;
     font-size: 64px;
     font-weight: 500;
     color: #444;
     cursor: pointer;
-  }
+}
 
-  #wrapper {
+#wrapper {
     display: flex;
     flex-direction: column;
     align-items: center;
-  }
+}
 
-  .toggle-container {
+.toggle-container {
     display: flex;
     justify-content: center;
     align-items: center;
     margin-bottom: 15px;
-  }
+}
 
-  .toggle-switch {
+.toggle-switch {
     position: relative;
     width: 50px;
     display: inline-block;
     vertical-align: middle;
-  }
+}
 
-  .toggle-switch input {
+.toggle-switch input {
     display: none;
-  }
+}
 
-  .toggle-switch-label {
+.toggle-switch-label {
     display: block;
     overflow: hidden;
     cursor: pointer;
@@ -69,9 +69,9 @@
     border-radius: 20px;
     background-color: #fff;
     height: 25px;
-  }
+}
 
-  .toggle-switch-switch {
+.toggle-switch-switch {
     display: block;
     width: 50%;
     margin: 0px;
@@ -83,15 +83,19 @@
     border-radius: 20px;
     transition: all 0.1s ease-in-out;
     height: 100%;
-  }
+}
 
-  .toggle-switch
-    input:checked
-    + .toggle-switch-label
-    .toggle-switch-switch {
+.toggle-switch
+input:checked
++ .toggle-switch-label
+.toggle-switch-switch {
     right: 0%;
-  }
+}
 
-  .toggle-text {
+.toggle-text {
     margin: 0 10px;
-  }
+}
+
+#names-sequence-block {
+    cursor: pointer;
+}
