+++ tests/app.test.js
@@ -1,0 +1,6 @@
+import { expect, test } from 'vitest'
+import {pickName} from "../src/app.js";
+
+test('True is truthy', () => {
+    expect(true).toBeTruthy();
+})
