+++ vite.config.js
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+    plugins: [],
+    test: {
+        environment: 'happy-dom'
+    }
+})
